打印内存

+++ npc/playground/cpp/common.h
@@ -6,7 +6,7 @@
 #include <string.h>
 #include <stdarg.h>
 
-typedef uint64_t word_t;
+typedef uint32_t word_t;
 typedef int64_t  sword_t;
 #define FMT_WORD "0x%016lx"
 #define FMT_PADDR "0x%016lx"
